requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
requests
pika
chromadb
pydantic>=2.0
